+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [#53] Add Security field to `dogu_v2`, which can be to define security policies for the dogu.
+  - This can be used for example in the pod security context on the kubernetes platform. 
 
 ## [v0.15.0] - 2024-11-13
 ### Added